Remove travis.yml and test_comprehensiveness.py
These are being removed as they do not currently provide any useful information as to the functioning of the Dashboard and takes far too long to run.
They were added as part of an initial project to increase Dashboard tests from zero but this has since been rescoped and determined to be a much larger refactoring project. Travis is being removed so it doesn't inhibit future work or provide a false sense of security when the non-useful test passes.