refactor: Generate file/path strings centrally
Code refactored so that any file/path that is needed is generated
using a central set of functions in config.py.  This has been done
as part of the refactor to Django so that the code can be moved
around and any changes in paths can be changed in just one place.