refactor: Move Dashboard source code into dedicated directory
Moved source code (not including code to fetch data), templates and
static content into a dedicated source code folder to accommodate the
transition to Django's directory structure.  To make the Dashboard
work the path constants were updated in config.py and some small
path-related changes to the tests were applied so that they run in the
new location.