refactor: Created views to render the top level pages.
This commit implements a URL structure for the Dashboard and includes
views for the top level navigation of the Dashboard as well as the
main pages in the Headlines section.  This required changes in the
templates to remove Flask-specific template functions and minor
syntax changes that prevented rendering.  The views themselves will
eventually benefit from some optimisation.  Note, this commit is a
breaking change for make_html.py and now the repo will not generate
a Flask website.