refactor: Renamed source code directory names for clarity
For clarity in the repository this commit renames the directories
that store the source code.  All the source code in src/ was moved
to dashboard/.  Also, the actual Django app code was moved from
src/dashboard/ to dashboard/ui/.  This commit doesn't change the
import statements to make the app run, this will be done in the next
commit.