refactor: Added exploring data/element(s) pages
Refactored the templates for the elements and element detail pages
in the exploring data section, completed the url routes and wrote
the view functions.  As part of this also found a bug due to changed
behaviour in xmlschema that required a change to is_valid_element,
fixed and renamed to is_valid_element_or_attribute to reflect its
behaviour for checking if a given path points to an element or an
attribute.