testing: Added basic automated testing of URL routes
Using the Django test framework this commit adds some basic automated
testing of the Dashboard URL routes, contains some fixes to the view
functions to properly return HTTP 404 errors, and updates the
documentation (including a small fix for the file paths).  A small fix
was also added to pytest.ini.