templates: Updated Boostrap CSS/JS links and added IATI design system
Removed <script> and <link> references to Bootstrap JS and full CSS
and replaced with links to Bootstrap grid and utility css to just
include functionality we want and not over-ride any functionality
from the design system.  Added IATI design system CSS and JS. We
will need to replace specific Boostrap components, but this step
is to avoid conflicts with the design system and to make sure what
is needed for the Dashboard to work.