templates: Replaced base footer with IATI design system template
This commit replaces the base footer with the standard footer from the
IATI design system.  As part of replacing the footer I also removed the
wrapper div.  Two of the custom styles are now redundant and so they
are also removed.  Note that this commit doesn't give a working
newsletter form, it just provides the form elements.