feat templates: replace bootstrap .panel with custom dashboard panel
Bootstrap .panel classes are not in the current version of Bootstrap so
this commit changes the Dashboad to use custom .dashboard-panel-* classes
in concert with .iati-card in preparation for new design system
components.  Some additional minor changes in layout and <h*> tags were
also applied as part of this change.