feat: updates to improve breadcrumb navigation
This commit fixes an issue where the breadcrumbs would not show the
correct location on "detail" pages, such as being on a publisher
detail page.  In this case the breadcrumb wouldn't show the publisher
name and the publishers link in the breadcrumb would be inactive.
This commit adds these breadcrumbs and corrects the links.  As part
of making this change, the breadcrumb navigation logic is moved from
the base.html template into the _make_context() function in
ui/views.py (also also the individual view functions for "detail"
pages.