templates: Move templates dir to jinja2, so automatic ADD_DIRS works
The django Jinja2 backend looks for a "jinja2" directory, not
"templates". By changing the directory we can use the automatic
"APP_DIRS" detection, instead of needing to specify it in "DIRS"
explicitly.

More importantly, we can enable "APP_DIRS" for the DjangoTemplates
backend, and it won't find our jinja2 templates, but will find the
Django Rest Framework default html templates. Therefore we can remove
the setting that enables only the JSON output.